Filter application in RTDs measures for PID algorithm
Variables storing RTDs values were redefined and vectors were included to maintain the last ten values of RTDs 2, 3, 5 and 6. The input of the PID algorithm is now stated as the average of the medians of the last ten values of the two corresponding RTDs (5 and 6 for heater 1, 2 and 3 for heater 2). Also, a verification was created to prevent errors in the last ten cycles, due to initialization of vector.